fix(demo): show user-friendly authentication errors
