+++ .changeset/friendlier-demo-error-banners.md
@@ -1,0 +1,9 @@
+---
+'ePDS': patch
+---
+
+Demo client's error banners now use plain user-friendly language instead of developer wording.
+
+**Affects:** End users
+
+**End users:** several of the demo's sign-in error banners used technical wording aimed at developers ("PAR rejected the request — check server logs", "token exchange failed") that didn't tell you anything actionable. The wording is now plain English and suggests retrying, contacting support if startup failures persist, or beginning sign-in again when the session has expired.
