+++ packages/demo/src/app/components/LoginForm.tsx
@@ -2,21 +2,8 @@
 
 import { useSearchParams } from 'next/navigation'
 import { useState } from 'react'
+import { getLoginErrorMessage } from '../../lib/login-errors'
 import { ForceLoginCheckbox } from './ForceLoginCheckbox'
-
-const ERROR_MESSAGES: Record<string, string> = {
-  auth_failed: 'Authentication failed. Please try again.',
-  session_expired:
-    'Your sign-in took too long to finish. Please sign in again.',
-  par_failed:
-    'Could not start login — the PDS rejected the request. Check server logs.',
-  invalid_email: 'Please enter a valid email address.',
-  invalid_handle: 'Please enter a valid handle (e.g. you.bsky.social).',
-  invalid_login_hint: 'Invalid login hint format.',
-  token_failed: 'Login could not be completed — token exchange failed.',
-  state_mismatch:
-    'Login session expired or was tampered with. Please try again.',
-}
 
 /**
  * Interactive login form (client component). Picks up theme colours via
@@ -25,9 +12,7 @@ const ERROR_MESSAGES: Record<string, string> = {
 export function LoginForm() {
   const searchParams = useSearchParams()
   const errorCode = searchParams.get('error')
-  const errorMessage = errorCode
-    ? ERROR_MESSAGES[errorCode] || `Unexpected error: ${errorCode}`
-    : null
+  const errorMessage = getLoginErrorMessage(errorCode)
   const [submitting, setSubmitting] = useState(false)
   const [mode, setMode] = useState<'email' | 'handle'>('email')
